examples/adk/requirements.txt
aioconsole == 0.8.2
coa-dev-coagent == 0.7.1
google-adk == 1.18.0
litellm == 1.79.1
openai == 2.6.1
python-dotenv == 1.2.1
